+++ nextstep-mvc/src/main/java/nextstep/mvc/adapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
-package nextstep.mvc.tobe.adapter;
+package nextstep.mvc.adapter;
 
-import nextstep.mvc.tobe.view.ModelAndView;
+import nextstep.mvc.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
