+++ nextstep-mvc/src/test/java/nextstep/mvc/handler/AnnotationHandlerMappingTest.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.handler;
+package nextstep.mvc.handler;
 
 import nextstep.db.DataBase;
 import nextstep.mvc.tobe.User;
