+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/TestUserController.java
@@ -1,6 +1,6 @@
 package nextstep.mvc.tobe;
 
-import nextstep.mvc.tobe.view.ModelAndView;
+import nextstep.mvc.view.ModelAndView;
 import nextstep.web.annotation.PathVariable;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
